pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openstack-tenantctl"
version = "0.4.0"
description = "Project-as-Code for OpenStack — declarative tenant provisioning enabling IaaS"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openstacksdk>=4.0,<5.0",
    "pyyaml>=6.0,<7",
    "deepmerge>=1.1,<3",
    "tenacity>=8.2,<10",
    "filelock>=3.12,<4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<10",
    "pytest-cov>=6.0,<8",
    "pytest-mock>=3.10,<4",
    "ruff>=0.15,<0.16",
    "black>=24.0,<26",
    "mypy>=1.15,<2",
    "types-requests>=2.31,<3",
    "types-PyYAML>=6.0,<7",
    "bandit[toml]>=1.7,<2",
    "build>=1.0,<2",
    "twine>=6.0,<7",
    "bump-my-version>=1.0,<2",
]

[project.scripts]
tenantctl = "src.main:cli"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    # --- Existing ---
    "E", "F", "W", "I", "N", "UP", "B", "SIM", "RUF",
    # --- Tier 1: low noise, high value ---
    "C4",    # flake8-comprehensions
    "PT",    # pytest style
    "RET",   # return style
    "PIE",   # misc simplifications
    "T20",   # print detection
    "TCH",   # type-checking imports
    "PERF",  # performance anti-patterns
    # --- Tier 2: guardrails with targeted ignores ---
    "S",     # bandit security
    "EM",    # exception message style
    "FBT",   # boolean trap
    "G",     # logging format
    "PTH",   # pathlib
]
ignore = [
    "PT019",  # fixture-param-without-value — false positives on factory fixtures
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101", "S108", "FBT001"]  # assert + /tmp paths + pytest parametrize bools are fine in tests
"src/main.py" = ["T201"]       # print() is intentional CLI output

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
    "error::openstack.warnings.RemovedInSDK50Warning",
]
